ecfr-analytics/ai_service/requirements.txt
google-cloud-aiplatform>=1.38.1
google-cloud-storage>=2.10.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.4.2
python-dotenv>=1.0.0
requests>=2.31.0